refactor: relocate backend overlay to backend/managed
Follow-up to #76 — the frontend overlay already lives at
frontend/managed/ because Turbopack refuses to resolve imports that
walk outside the Next.js project root. Moving the backend overlay to
backend/managed/ restores the symmetry: each sub-project owns its own
managed/ subdirectory, with no stray managed/ dir at the repo root.

- managed/backend/ → backend/managed/
- managed/backend/alembic.ini script_location updated
- start.sh points at backend/managed/alembic.ini
- backend/main.py imports from backend.managed.auth0
- managed/README.md → MANAGED.md (repo root) and updated to explain
  the per-subproject layout and why it isn't a single top-level dir
