+++ backend/managed/alembic.ini
@@ -1,5 +1,5 @@
 [alembic]
-script_location = managed/backend/migrations
+script_location = backend/managed/migrations
 prepend_sys_path = .
 version_path_separator = os
